refactor: getTodayInfo, getTotalCalendarCellsInfo 유틸함수 별도 파일로 분리 및 이에따른 경로 수정
