+++ src/app/(main)/page.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@mui/material'
 import LogoutButton from '@/shared/components/LogoutButton'
-import { getTodayInfo } from '@/shared/utils/getCalendarInfo'
+import getTodayInfo from '@/shared/utils/getTodayInfo'
 import { Month } from '@/shared/types/type'
 import CalendarControlUI from './_component/CalendarControlUI'
 import CalendarUI from './_component/CalendarUI'
@@ -19,7 +19,6 @@ import UserName from './_component/UserName'
 // todo: 지역적인 utils은 해당 폴더에 위치하도록 변경
 // todo: 나중에 타입도 중복없도록 정리하기
 
-// todo : servercomponent 에서 params 받는 방법도 있음
 // todo: 성능 테스트 (light house)
 // todo: test tool 로 시나리오 테스트 (jest)
 // todo: test
@@ -27,12 +26,6 @@ import UserName from './_component/UserName'
 // todo: 클라이언트 컴포넌트 로딩 처리
 
 export default function MainPage({ searchParams }: { searchParams: { [key: string]: string | string[] | undefined } }) {
-  // 직접 가져왔으면 어땟을까?
-  // context에 저장하지 말고 localStorage에 저장 하는게 나을 듯 하다.
-  // 그리고 굳이 middleware에서 cookie 안써도 될 듯 싶고 -> mi
-  // 그리고 굳이 isLoading 필요 없을 듯 싶다.
-  // const context = useContext(AuthContext)
-
   const todayInfo = getTodayInfo()
   const [selectedYear, selectedMonth] = [
     Number(searchParams.year) || todayInfo.year,
